+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog 
 
+## Unreleased
+
+- breaking change: `dc_update_device_chats()` was removed. This is now done automatically during configure.
+
+- Added a `bot` config. Currently, it only prevents filling the device chats automatically.
+
 ## 1.46.0
 
 - breaking change: `dc_configure()` report errors in